feat: change gitignore, main.py updated del: main_test.py
