+++ .gitignore
@@ -3,3 +3,4 @@
 *.xml
 *.pyc
 __pycache__/
+*test*
